Remove all use of the nfsctl system call.
This systemcall was deprecated early in the 2.6 series
as it was replaced by an in-kernel cache which was refilled
using an upcall.  All communication to kernel is now through
the nfsd filesystem.

The nfsctl systemcall itself was removed in 3.1.

It is unlikely to have been used for over a decade.

To remove all uses for the nfsctl systemcall, and call code that only
runs when "new_cache" is false.  We now assume "new_cache" is always
true.

This allows the removal of several files as well as assorted functions.

Signed-off-by: NeilBrown <neilb@suse.com>
Signed-off-by: Steve Dickson <steved@redhat.com>