conffile: free image of config file after parsing
conffile allocates memory and reads in the config file.
It then parses the file, using strdup() to take a copy of
any string that it uses, so after conf_parse() there are no
references in to the allocated file image.

conffile does not free this image.  It keeps a pointer, but never uses
it in an interesing way, and never frees it.  This is a little clumsy
and interfers with a future patch which will support the inclusion of
subordinate config files.

So free 'new_conf_addr' when finished with it, and discard the
'conf_addr' variable that stored it.
This has an insignificant performance consequence in that we node
always free everything in the hash table, even when we know it must
be empty.

Signed-off-by: NeilBrown <neilb@suse.com>
Signed-off-by: Steve Dickson <steved@redhat.com>