conffile: allow $name expansion of tag values.
If the value for a tag starts with '$', then the remainder
of the value is treated as an environment variable name.
It is looked up in the environment (getenv) and if not found,
it is looked for in the [environment] section of the config file.

This lookup is formed as access time e.g. by conf_get_str(), not at
parse time.

The expected usage is that the config file can contain something like

 [environment]
  include = /etc/sysconfig/nfs
 [other-section]
  tag = $NAME

and conf_get_str("other-section","tag") will report the value of "NAME"
in the given file.

As different distributions used different environment files, and
different
variable names with-in them, a distro could provide a static config file
which maps from names in that environment file to config tags requires
by NFS daemons.

Signed-off-by: NeilBrown <neilb@suse.com>
Signed-off-by: Steve Dickson <steved@redhat.com>