nfsd: move and improve test on valid port
nfssvc_set_sockets() access textual port numbers (by lookup in
/etc/services).  This uses getaddrinfo which reports errors, except
for out-of-range numbers.

So change the test on a valid port to only complain if the port given
is purely numeric, but is out-of-range.

Also move it so that any default value gets tested the same as any
argument value.

Signed-off-by: NeilBrown <neilb@suse.com>
Signed-off-by: Steve Dickson <steved@redhat.com>