umount.nfs: assume path name is canonical.
/usr/bin/umount will always pass a canonical name
to umount.nfs, so it is safe to disable canonicalization.

When umounting an NFS filesystem, it is generally safest to
not "stat" the mountpoint at all as that can block
indefinitely.  umount() will not block, but lstat() etc can.
By disabling canonicalization in libmount, we discourage it
from ever calling 'stat' family operations, and thus reduce
the chance of a hang.

Note that to be fully effective, this requires changes to
util-linux which have not yet been accepted.
When both that change and this are in effect, automounters
can use "umount -c $PATH" to safely unmount a filesystem
without blocking.

Signed-off-by: NeilBrown <neilb@suse.com>
Signed-off-by: Steve Dickson <steved@redhat.com>